init: Change how the global seller is initialized
Instead of trusting the fact that the newly generated list of sellers
matches the order of the existing one, perform a look up to find the
real position of the new seller in the global array, and do not mix
usage of local and global array.

Vendors do not need this kind of treatment because there are no
markets, and the local and global list always match.