auth: Use thepassed in headers instead of coming up with hosts
With multiple hosts spawning (lorcana, magic) and their aliases (www, magic)
there is a chance that a signature is made invalid depending on the
visited host. Instead, make sure that there is a valid default value
mimicking the domain option of cookies. The changes could be more
thorough, but in this way it minimizes disruption to existing signatures.